[IMP] #9242 Survive broken 'git am' patches
We recently started using a new way of applying MR's to branches.
Instead of letting gitaggregator merge, we apply a shell_command_after
that applies the MR as a 'git am' patch, which preserves pinning.
However when 'git am' fails, the directory is left in a broken state
and this should be fixed upon rebuild.